Make gts work with workspaces
I noticed that gts fix and gts link wasn't working in the npm workspaces. This is related to the issues:

gts monorepo does not work with npm ? google/gts#752
gts does not work in a yarn workspaces monorepo google/gts#490
The last issue is with yarn but has more discussion. One solution was using nohoist. This is not a default option with npm. So I tried with the following package https://github.com/zgriesinger/noist. But this did not fix the errors. The last solution mentioned was only linting from the root level; this works.

I have also subscribed to the issue [feature] support npm workspaces. There may be a better solution in the future.